Constructors
Contructors : A constructor in Java is a special method that is used to initialize objects. The constructor is called when an object of a class is created. It can be used to set initial values for object attributes.
Contructor parameter : Constructors can also take parameters, which is used to initialize attributes.