Iterator
An Iterator is an object that can be used to loop through collections. It is called an "iterator" because "iterating" is the technical term for looping.
The iterator() method can be used to get an Iterator for any collection.
To loop through a collection, use the hasNext() and next() methods of the Iterator.
Iterators are designed to easily change the collections that they loop through. The remove() method can remove items from a collection while looping.