Exceptions
Throw an exception if age is below 18 (print "Access denied"). If age is 18 or older, print "Access granted".
If an error occurs, we can use try...catch to catch the error and execute some code to handle it.
The throw statement allows you to create a custom error.
The throw statement is used together with an exception type. There are many exception types available in Java: ArithmeticException, FileNotFoundException, ArrayIndexOutOfBoundsException, SecurityException, etc.